+++ docs/changelog.md
@@ -4,6 +4,7 @@
 ## New
 - By request, added pizza song social
 - Added yodel sound by request.
+- Pointed a bunch of aliases for vomit to play the vomit sound.
 
 ### Fixed
 - Updated tradesman marketplace to support names like McKnight so sounds would play. Also updated the display to match the new rejex patterns, and tightened things up just a bit. I tested most things but stuff might be odd so if something is, please post on metaf.